[FIX] Failures in early logging due to static object destruction
In CString::Add, a static std::regex object is used to parse the
provided format string, but if failures occur in early initialization
code, exiting threads will cause the object to be destructed and
subsequent prints will fail.

Fix this by introducing a wrapper class that allows us to prevent the
destructor from being called on the regex object so its lifetime matches
that of the entire program.